requirements.txt
# Core Dependencies
langchain==1.0.1
langchain-openai==1.0.1
langchain-community==0.4.0

# embeddings
langchain-huggingface==1.0.1
sentence-transformers==3.0.1

# dependency injector
dependency-injector==4.41.0

# Azure and Fabric
azure-identity==1.15.0
azure-appconfiguration==1.7.2
azure-keyvault-secrets==4.7.0
semantic-link==0.6.0

# MCP and API clients
mcp==1.19.0
requests==2.32.5
httpx==0.27.1

# Data handling
pydantic==2.12.2
pydantic-settings==2.11.0
numpy==1.26.4

# CLI
click==8.1.7
rich==13.7.1

# Logging and monitoring
structlog==24.1.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1